Move path_utils to common/cpp for use in other embeddings, add unit test
