+++ shell/platform/common/cpp/BUILD.gn
@@ -2,6 +2,8 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
+import("//flutter/testing/testing.gni")
+
 config("desktop_library_implementation") {
   defines = [ "FLUTTER_DESKTOP_LIBRARY" ]
 }
@@ -44,6 +46,7 @@ source_set("common_cpp") {
   configs += [ ":desktop_library_implementation" ]
 
   deps = [
+    ":common_cpp_core",
     ":common_cpp_library_headers",
     "//flutter/shell/platform/common/cpp/client_wrapper:client_wrapper",
     "//flutter/shell/platform/embedder:embedder_with_symbol_prefix",
@@ -58,6 +61,41 @@ source_set("common_cpp") {
     # For wstring_conversion. See issue #50053.
     defines += [ "_SILENCE_CXX17_CODECVT_HEADER_DEPRECATION_WARNING" ]
   }
+}
+
+# The portion of common_cpp that has no dependencies on the public/
+# headers. This division should be revisited once the Linux GTK
+# embedding is futher along and it's clearer how much, if any, shared
+# API surface there will be.
+source_set("common_cpp_core") {
+  public = [
+    "path_utils.h",
+  ]
+
+  sources = [
+    "path_utils.cc",
+  ]
+}
+
+test_fixtures("common_cpp_core_fixtures") {
+  fixtures = []
+}
+
+executable("common_cpp_core_unittests") {
+  testonly = true
+
+  sources = [
+    "path_utils_unittests.cc",
+  ]
+
+  deps = [
+    ":common_cpp_core",
+    ":common_cpp_core_fixtures",
+    "//flutter/testing",
+    "//third_party/dart/runtime:libdart_jit",
+  ]
+
+  public_configs = [ "//flutter:config" ]
 }
 
 copy("publish_headers") {
