+++ shell/platform/common/cpp/path_utils.h
