+++ shell/platform/glfw/BUILD.gn
@@ -36,8 +36,6 @@ source_set("flutter_glfw") {
     "key_event_handler.cc",
     "key_event_handler.h",
     "keyboard_hook_handler.h",
-    "path_utils.cc",
-    "path_utils.h",
     "platform_handler.cc",
     "platform_handler.h",
     "text_input_plugin.cc",
